requirements.txt
gradio
sentence-transformers
pillow
pandas
